backend/database-engineering/cursor/pyproject.toml
[project]
name = "cursor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary>=2.9.11",
]
